Rename pkg
Use space for indent, fix paths
