+++ Glaide.jl/app/time_dependent_inversion.jl
@@ -1,4 +1,4 @@
-using Reicalg
+using Glaide
 using CUDA
 using CairoMakie
 using Printf
