+++ Glaide.jl/src/snapshot/model.jl
